+++ smap_io/__init__.py
@@ -1,7 +1,8 @@
-# -*- coding: utf-8 -*-
 import pkg_resources
 
 try:
     __version__ = pkg_resources.get_distribution(__name__).version
 except:
     __version__ = 'unknown'
+
+from smap_io.interface import *