+++ CHANGES.rst
@@ -2,9 +2,25 @@
 Changelog
 =========
 
+Version 0.X
+===========
+
+- Add kwargs to time series reader
+- Add option for download checking
+- Add CRID reading
+- Name PM variables *_pm in time series
+- Add download module
+- Add SMAP L3 v4 and v5 support
+- Update readme
+
+Version 0.2
+===========
+
+- Add metadata from netCDF file to returned image.
+- Add option to return data as 1D arrays.
+- Add image to time series conversion and time series reading interface.
+
 Version 0.1
 ===========
 
-- Feature A added
-- FIX: nasty bug #1729 fixed
-- add your changes here!
+- Initial version with one dataset supported.